starting to clean up types
