+++ db/transformations.test.ts
@@ -1,6 +1,7 @@
 // Run via `yarn test db/Row.test.ts`
 
-import { getFinalCsvRow } from './Row';
+import { CsvRow } from '..';
+import { getFinalCsvRow } from './transformations';
 
 // eslint-disable-next-line max-lines-per-function
 describe('Row', () => {
@@ -12,8 +13,9 @@ describe('Row', () => {
     const ftCurrencyIn = 'USDC';
     const args = {
       args_base64: 'eyJhbW91bnQiOiIxMDAwMDAwMDAwMCIsInJlY2VpdmVyX2lkIjoiZGV2M2lzLm5lYXIifQ==',
-      args_json: undefined,
+      args_json: '{"amount":"10000000000","receiver_id":"dev3is.near"}',
       deposit: 1,
+      gas: 1,
       method_name: 'ft_transfer',
     };
     const indexerRow = {
@@ -22,13 +24,13 @@ describe('Row', () => {
       args,
       block_height: 1,
       // eslint-disable-next-line @typescript-eslint/no-loss-of-precision
-      block_timestamp: 1_659_973_934_799_291_015, // TODO Fix this in TTA-63
+      block_timestamp: 1_659_973_934_799_291_015,
       receipt_predecessor_account_id: 'nf-payments2.near',
       receipt_receiver_account_id: accountId,
       receiver_account_id: accountId,
       transaction_hash: '6xnB3tApHAhRjStzhfRHRz5Z2dAWZZnkVgCZoWLypcsT',
     };
-    const expectedResult = {
+    const expectedResult: CsvRow = {
       /* eslint-disable canonical/sort-keys */
       date: '2022-08-08',
       account_id: accountId,
@@ -36,7 +38,7 @@ describe('Row', () => {
       block_timestamp: indexerRow.block_timestamp,
       from_account: indexerRow.receipt_predecessor_account_id,
       block_height: indexerRow.block_height,
-      args: '{"amount":"10000000000","receiver_id":"dev3is.near"}',
+      args_json: '{"amount":"10000000000","receiver_id":"dev3is.near"}',
       transaction_hash: indexerRow.transaction_hash,
       amount_transferred: nearAmount,
       // Fungible Token
